% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_msoa_names.R
\name{add_msoa_names}
\alias{add_msoa_names}
\title{Add the House of Commons Library MSOA Names to a data frame}
\usage{
add_msoa_names(df, welsh = TRUE)
}
\arguments{
\item{df}{A data frame. Obv.}

\item{welsh}{Whether to keep the Welsh language names. Defaults to TRUE.}
}
\value{
A wider data frame.
}
\description{
This will only work if you have a column in the df that matches
a (clean) column name from the MSOA Names resource (so the left_join works).
}
