% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_year.R
\name{as_year}
\alias{as_year}
\title{as_year}
\usage{
as_year(x, months = 1:12, unlist = FALSE)
}
\arguments{
\item{x}{a vector of years, provided as numeric or string}

\item{months}{a numeric vector. 1:12 by default (whole year).}

\item{unlist}{FALSE by default (allows returning a list). TRUE causes it to return a vector of dates.}
}
\value{
a vector of dates, or a list of vectors, one for each year, where multiple years are requested.
}
\description{
as_year
}
\examples{
head(as_year(2019:2020, months = c(1, 4), unlist = TRUE), 10)
}
