% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_it_simple.R
\name{batch_it_simple}
\alias{batch_it_simple}
\title{Convert a list or vector to a batched list of its elements}
\usage{
batch_it_simple(x, batch_size)
}
\arguments{
\item{x}{a vector, or a list flattenable to a vector}
}
\description{
Convert a list or vector to a batched list of its elements
}
\examples{
batch_it_simple(letters, 6)
batch_it_simple(letters, 0.45)
}
