% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mascot.R
\name{chars}
\alias{chars}
\title{Alias for stringr::str_split_1(..., pattern = "")
Split a string into its characters}
\usage{
chars(x)
}
\arguments{
\item{x}{A single character string}
}
\description{
Alias for stringr::str_split_1(..., pattern = "")
Split a string into its characters
}
