% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{Converts R colours to hex codes}
\usage{
col2hex(x)
}
\arguments{
\item{x}{character A vector of R colour names from base::colours()}
}
\description{
Adapted from https://github.com/talgalili/gplots/blob/master/R/col2hex.R
}
\examples{
col2hex("seagreen4")
}
