% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonest.R
\name{commonest}
\alias{commonest}
\title{Pull out the commonest (mode) item from a vector}
\usage{
commonest(vec, first = TRUE)
}
\arguments{
\item{vec}{A vector.}

\item{first}{If there is a tie, this parameter decides whether to prioritise the
element that appears first in `vec`.
If `FALSE`, the function will just return the first item in the result produced by
the default mode of `vec_count` (ie `"count"`, which does not return results
prioritised by location -- see test-commonest file).}
}
\value{
A single element of the same type as `vec`.
}
\description{
Pull out the commonest (mode) item from a vector
}
\examples{
set <- sample(letters, 20, replace = TRUE)
commonest(set)
commonest(set, first = FALSE)

test_df <- dplyr::tibble(letters = set)
dplyr::summarise(test_df, letter = commonest(letters))
}
