% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_rstudio_recent_projects.R
\name{edit_rstudio_recent_projects}
\alias{edit_rstudio_recent_projects}
\title{Provides easy edit access within RStudio to recent projects list
Windows only}
\usage{
edit_rstudio_recent_projects()
}
\description{
Provides easy edit access within RStudio to recent projects list
Windows only
}
