% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_sf_bbox.R
\name{expand_sf_bbox}
\alias{expand_sf_bbox}
\title{Create an expanded sf object}
\usage{
expand_sf_bbox(x, factor, crs = NULL)
}
\arguments{
\item{x}{A rectangular sf object, or any object with a bbox}

\item{factor}{The factor(s) to expand the bbox by. Can be length 1, 2 or 4.}

\item{crs}{the CRS of the returned object. The same as x by default. Use NA if no CRS is desired.}
}
\value{
a rectangular sf object
}
\description{
Create an expanded sf object
}
