% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_col_types.R
\name{extract_col_types}
\alias{extract_col_types}
\title{Extracts all the column types of a data frame to a character vector}
\usage{
extract_col_types(.data, named = FALSE, collapse = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{named}{logical: Whether to retain the column names as names for the vector of types. Default \code{FALSE}.}

\item{collapse}{logical: Whether to collapse the vector of column names to a single string comprising the initial letters of each type. Default \code{FALSE}.}
}
\value{
A character vector (with or without names), or if \code{collapse} is \code{TRUE}, a single condensed string.
}
\description{
Extracts all the column types of a data frame to a character vector
}
