% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_poly.R
\name{fit_square}
\alias{fit_square}
\title{Create a square box around a polygon}
\usage{
fit_square(x, return = c("bbox", "polygon"))
}
\arguments{
\item{x}{A polygon}

\item{return}{Whether to return just a bbox, or an sf polygon}
}
\description{
Create a square box around a polygon
}
