% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stats.R
\name{generate_stats}
\alias{generate_stats}
\title{Generate stats tables or summaries of age-based data by area}
\usage{
generate_stats(
  area,
  measure,
  source_pcts,
  invert = TRUE,
  age_bands = "lloyds_all",
  summary = TRUE
)
}
\arguments{
\item{area}{name of geographical area}

\item{measure}{name of data measure}

\item{source_pcts}{age band percentages}

\item{invert}{whether to convert percentages into their inverse (100-x). Default TRUE}

\item{age_bands}{use an alias for certain common age bandings (lloyds_all, lloyds_wa, lloyds_65, ons, ofcom), or supply a list of two integer vectors, named `lows` and `highs`}

\item{summary}{whether to output data for all age bands (FALSE) or just a total for the area (TRUE, default)}
}
\description{
Generate stats tables or summaries of age-based data by area
}
