% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf.R
\name{get_native_list_separator}
\alias{get_native_list_separator}
\title{return the list separator character as set up in your system locale}
\usage{
get_native_list_separator()
}
\value{
a character ("," or ";" or ...)
}
\description{
return the list separator character as set up in your system locale
}
