% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_data.R
\name{get_postcode_data}
\alias{get_postcode_data}
\title{Use postcodes.io To Get Postcode Metadata}
\usage{
get_postcode_data(postcodes)
}
\arguments{
\item{postcodes}{one or more valid GB postcodes}
}
\value{
Tibble of postcode data
}
\description{
Use postcodes.io To Get Postcode Metadata
}
\examples{
get_postcode_data("W1A 1AA")

}
