% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji}
\alias{gitmoji}
\title{Writes a gitmoji to the clipboard}
\usage{
gitmoji(x, exact = TRUE)
}
\arguments{
\item{x}{string. A piece of text to search for.}

\item{exact}{Boolean, default \code{TRUE}. Whether to only return a result for the
exact string provided, or return all names that match the string.}
}
\value{
\code{TRUE} if successful; the gitmoji is/are copied to the clipboard.
}
\description{
Use \code{\link[=gitmoji_name]{gitmoji_name()}} to list all available gitmoji or to search the table of
gitmoji by name.
Use \code{\link[=gitmoji_desc]{gitmoji_desc()}} to search the table of gitmoji by description
}
\examples{
gitmoji("memo")
gitmoji("flag")
}
