% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji_desc}
\alias{gitmoji_desc}
\title{List gitmoji whose description matches `x`
If `x` is "" (as by default), all rows in the table will be returned}
\usage{
gitmoji_desc(x = "")
}
\arguments{
\item{x}{string. A piece of text to search for}
}
\value{
a three-column table showing matching emoji with their code and description
}
\description{
List gitmoji whose description matches `x`
If `x` is "" (as by default), all rows in the table will be returned
}
