% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji_list}
\alias{gitmoji_list}
\title{List gitmoji names matching a text string (\code{x})
If \code{x} is "" (as by default), all names in the table will be returned}
\usage{
gitmoji_list(x = "", exact = FALSE)
}
\arguments{
\item{x}{string. A piece of text to search for}

\item{exact}{boolean, default TRUE. Whether to only return a result for the exact string provided, or return all names that contain the string.}
}
\value{
A character vector of all matching gitmoji names
}
\description{
List gitmoji names matching a text string (\code{x})
If \code{x} is "" (as by default), all names in the table will be returned
}
