% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji_name}
\alias{gitmoji_name}
\title{List gitmoji names matching \code{x}}
\usage{
gitmoji_name(x = "", exact = FALSE)
}
\arguments{
\item{x}{string. A piece of text to search for.}

\item{exact}{Boolean, default \code{TRUE}. Whether to only return a result for the
exact string provided, or return all names that match the string.}
}
\value{
A character vector of all matching gitmoji names
}
\description{
If \code{x} is "" (as by default), all names in the table will be returned
}
