% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_msoa_names.R
\name{hcl_msoa_names}
\alias{hcl_msoa_names}
\title{House of Commons Library MSOA Names as an easy tibble}
\usage{
hcl_msoa_names(year = c("2021", "2011"), keep_welsh = TRUE)
}
\arguments{
\item{year}{character. Whether to return names and codes for 2021 or 2011
MSOAs. "2021" is the default.}

\item{keep_welsh}{Boolean. Whether to retain Welsh language names (nmw
columns). \code{TRUE} by default.}
}
\value{
A tibble
}
\description{
House of Commons Library MSOA Names as an easy tibble
}
