% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_old_am_i.R
\name{how_old_am_i}
\alias{how_old_am_i}
\title{A quick function using `lubridate` to report how old I am}
\usage{
how_old_am_i(date = "1977-02-17")
}
\arguments{
\item{date}{your date of birth. Set to mine by default.}
}
\description{
A quick function using `lubridate` to report how old I am
}
