% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds.R
\name{import_rds}
\alias{import_rds}
\title{Find and Import RDS Files}
\usage{
import_rds(folder = "", extension = "Rds", recursive = TRUE)
}
\arguments{
\item{folder}{a sub-folder to only search within (default is "" ie working directory)}

\item{extension}{a filename extension to import (single values only)}

\item{recursive}{whether to search through folders recursively (default is TRUE)}
}
\value{
object(s) in the global env.
}
\description{
A quick way to read a load of Rds files into the global env.
The folder to search within can be specified, otherwise it will find all files
within the root/working directory (recursively, by default)
}
\examples{
import_rds("rds_data")
}
