% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAssPipe.R
\name{insertAssPipe}
\alias{insertAssPipe}
\title{Insert magrittr assignment pipe}
\usage{
insertAssPipe()
}
\value{
None

dont#@export
}
\description{
RStudio API tweak to insert the reflexive assignment pipe
from magrittr at the cursor location when editing R code.
Designed to be assigned a keyboard shortcut.
- I use Ctrl+Alt+,
}
