% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to_html5.R
\name{knit_to_html5}
\alias{knit_to_html5}
\title{Custom HTML5 knit function}
\usage{
knit_to_html5(input, ...)
}
\arguments{
\item{input}{input document to be knitted}

\item{...}{other arguments to pass to rmarkdown::output_format()}
}
\description{
Custom HTML5 knit function
}
