% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{my_prompt}
\alias{my_prompt}
\title{My custom prompt}
\usage{
my_prompt(unicode = FALSE, rstd_col = "darkseagreen1")
}
\arguments{
\item{unicode}{whether to use unicode characters in the prompt.}

\item{rstd_col}{colour to use for the RStudio codename}
}
\description{
My custom prompt
}
