% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rmd.R
\name{new_rmd}
\alias{new_rmd}
\title{Create a new RMarkdown draft based on a template}
\usage{
new_rmd(
  file = "index.Rmd",
  template = "myrmidon2",
  package = "myrmidon",
  create_dir = FALSE,
  edit = TRUE
)
}
\arguments{
\item{file}{File name for the draft}

\item{template}{Template to use as the basis for the draft. This is either
the full path to a template directory or the name of a template directory
within the \code{rmarkdown/templates} directory of a package.}

\item{package}{(Optional) Name of package where the template is located.}

\item{create_dir}{\code{TRUE} to create a new directory for the document
(the "default" setting leaves this behavior up to the creator of the
template).}

\item{edit}{\code{TRUE} to edit the template immediately}
}
\description{
Create a new RMarkdown draft based on a template
}
