% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_age_profiles.R
\docType{data}
\name{ons_age_profiles}
\alias{ons_age_profiles}
\title{ONS 2020 Age Estimates by single year of age for UK administrative areas}
\format{
420 rows, 95 variables (cols 2, `all_ages`, gives the total for all ages; cols 5-95 are the single year of age populations from age_0 to age_90 (includes 91+))
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatesforukenglandandwalesscotlandandnorthernireland}
}
\usage{
ons_age_profiles
}
\description{
ONS 2020 Age Estimates by single year of age for UK administrative areas
}
\keyword{datasets}
