% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationalise_df.R
\name{rationalise_df}
\alias{rationalise_df}
\title{Shuffle the columns of a data frame in decreasing order of the number of unique values in each column}
\usage{
rationalise_df(.data)
}
\arguments{
\item{.data}{A data frame}
}
\description{
This makes sense when you have "nested" values such as sub-categories of
categories and you want the top-level categories at the right-hand end
of your data frame and the more fine-grained or unique values at the
left-hand end. If you have been doing left joins you might end up with an
"irrationally sorted" table; this function will help you rationalise it.
This won't make sense for all datasets, naturally.
}
