% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rss.R
\name{read_rss}
\alias{read_rss}
\title{Read an RSS feed and return items data and feed metadata}
\usage{
read_rss(url)
}
\arguments{
\item{url}{URL of the RSS feed}
}
\description{
Read an RSS feed and return items data and feed metadata
}
