% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_pdf_font.R
\name{register_pdf_font}
\alias{register_pdf_font}
\title{Register a font for use in PDF creation}
\usage{
register_pdf_font(name)
}
\arguments{
\item{name}{character. The name of the font (or the font filename) you
want to register}
}
\description{
Based on work by Konrad Rudolph here
https://github.com/klmr/ggplots/blob/main/fonts.r
}
