% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex_to_gist.R
\name{reprex_to_gist}
\alias{reprex_to_gist}
\title{Easily post reprexes from selected text to a GitHub gist}
\usage{
reprex_to_gist()
}
\description{
If no selected text is found for the reprex input then the currently visible
open file, and then the system clipboard, will be tried in that order.
}
\details{
From https://github.com/tidyverse/reprex/issues/190#issuecomment-817313938
by Mickaël Canouil
}
