% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_it.R
\name{save_it}
\alias{save_it}
\title{Shortcut to Save an Object as an RDS File}
\usage{
save_it(x, dir = "rds_data")
}
\arguments{
\item{x}{Object to save (can be piped in)}

\item{dir}{Directory to store RDS files ('rds_data' by default)}
}
\value{
Creates an RDS file
}
\description{
Shortcut to Save an Object as an RDS File
}
