% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_data_join.R
\name{suggest_postcode_fixes}
\alias{suggest_postcode_fixes}
\title{Return data for suggested replacements for invalid postcodes}
\usage{
suggest_postcode_fixes(codes)
}
\arguments{
\item{codes}{character vector. Postcodes that may be invalid.}
}
\value{
A data frame with a row for each invalid postcode supplied.
}
\description{
Data for terminated invalid codes will be returned according to the nearest
current valid code for the old code's longitude/latitude.
Data for codes that are invalid due to an incorrect final letter will be
returned according to an autocompletion process.
Missing data will be returned for any other invalid postcodes that could not
be replaced using the two methods above.
Any valid codes supplied will be ignored, and excluded from the returned data
}
\examples{
suggest_postcode_fixes(c("hd1 2ut", "hd1 2uu", "hd1 2uv"))
}
