% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{switch_my_prompt}
\alias{switch_my_prompt}
\title{Toggle unicode usage on/off in my custom prompt `my_prompt()`}
\usage{
switch_my_prompt(...)
}
\description{
Toggle unicode usage on/off in my custom prompt `my_prompt()`
}
