% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_linen.R
\name{theme_clay}
\alias{theme_clay}
\title{A ggplot2 theme specifically for maps, with a dark ("ebony clay") background}
\usage{
theme_clay(font_family = "Fira Sans")
}
\arguments{
\item{font_family}{Specify a font family to use. "Fira Sans" by default.}
}
\description{
Starts with theme_void() and removes all gridlines and axis markings and
labels. Legend moves inside the panel area and is transparent.
}
