% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_linen.R
\name{theme_linen}
\alias{theme_linen}
\title{My preferred ggplot2 theme, based on a "linen" coloured background}
\usage{
theme_linen(font_family = "Fira Sans")
}
\arguments{
\item{font_family}{Specify a font family to use. "Fira Sans" by default.}
}
\description{
My preferred ggplot2 theme, based on a "linen" coloured background
}
