% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_tbl_wider.R
\name{transpose_tbl_wider}
\alias{transpose_tbl_wider}
\title{Neatly transpose a long tibble to wide format}
\usage{
transpose_tbl_wider(.tbl)
}
\arguments{
\item{.tbl}{A tibble or data frame}
}
\description{
The first column will be used to provide the column headings for the new
columns
}
