% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrmidon-package.R
\name{view}
\alias{view}
\title{Borrow \code{view()} from \code{{tibble}}}
\usage{
view(x)
}
\arguments{
\item{x}{a data frame or similar viewable object}
}
\description{
Borrow \code{view()} from \code{{tibble}}
}
