% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xl.R
\name{view_xl}
\alias{view_xl}
\title{View a data frame in Excel}
\usage{
view_xl(.data)
}
\arguments{
\item{.data}{A data frame or tibble}
}
\value{
Passes \code{.data} through, invisibly (so can be used in a pipeline?)
}
\description{
View a data frame in Excel
}
