% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf.R
\name{wtf}
\alias{wtf}
\title{open a data.frame in excel}
\usage{
wtf(x)
}
\arguments{
\item{x}{a data.frame to open in excel}
}
\value{
nothing
}
\description{
wtf (short for writetempfile) will write a data.frame to a temp csv file and open it in excel
}
