% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf.R
\name{wtx}
\alias{wtx}
\title{open data.frame in temporary excel file}
\usage{
wtx(x, path = dirname(tempfile()), fn = NULL, sheet = NULL, open_file = TRUE)
}
\arguments{
\item{x}{a data.frame to open in excel}

\item{path}{folder path where the file will be saved (default = temp folder)}

\item{fn}{file name (default = sheet name)}

\item{sheet}{sheet name (default = variable name)}

\item{open_file}{TRUE/FALSE to open EXCEL after writing file}
}
\value{
nothing
}
\description{
create a temp xlsx file with autofilter from a data.frame and open it in excel
}
