% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_dates.R
\name{year_dates}
\alias{year_dates}
\title{Create vectors of dates for one or more years (and/or months)}
\usage{
year_dates(x, months = 1:12)
}
\arguments{
\item{x}{A vector of years, provided as numeric or string.}

\item{months}{A vector of month numbers, provided as numeric or string. 1:12
by default (i.e. the whole of the year).}
}
\value{
A list of vectors of dates, one element per year requested.
}
\description{
Create vectors of dates for one or more years (and/or months)
}
\examples{
head(year_dates(2019:2020, months = c(1, 4)), 10)
}
