% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_it.R
\name{batch_it}
\alias{batch_it}
\title{Turn A Long Vector Into A Batched List}
\usage{
batch_it(x, batches = NULL, proportion = NULL, maximise = FALSE, quiet = TRUE)
}
\arguments{
\item{x}{a vector}

\item{batches}{size of batches to create. Can be a single value or
multiple values (see examples). Should be a whole, positive
number.}

\item{proportion}{proportional sizes of batches to be created, for
example c(4, 6) will create two batches of approximately 40% and
60% of the length of the target vector. When multiple proportions
are provided, these are not repeated. A single proportion less
than 1 is repeated as many times as possible to get near to the
length of the target vector. For example, a proportion of 0.1
will be treated as a tenth, and batch sizes will be rounded to an
integer size nearest to a tenth of the vector length.}

\item{maximise}{FALSE by default. When TRUE, a vector of batch
sizes will be partially repeated to fit maximally to the length
of the target vector. See examples below.}

\item{quiet}{Boolean, default TRUE. Whether to show ui_* messages.}
}
\description{
Go \code{batch_it()} crazy!
}
\details{
Batch up a long vector or list of vectors for passing to
services with limited APIs such as postcodes.io
}
\examples{
batch_it(seq(2, 60, 2), 6)
batch_it(seq(2, 60, 2), proportion = 0.2)
batch_it(1:100, batches = c(20, 30, 50))
batch_it(letters, batches = c(4, 6))
batch_it(letters, batches = c(4, 6), maximise = TRUE)
batch_it(letters, proportion = c(4, 6))

as_year <- function(x) {
  lubridate::as_date(
    lubridate::ymd(paste0(x, "-01-01")):
    lubridate::ymd(paste0(x, "-12-31"))
  )
}
month_lengths <- function(year) {
  lubridate::as_date(paste0(year, "-", 1:12, "-01")) \%>\%
    lubridate::days_in_month() \%>\%
    split(factor(names(.), levels = month.abb)) \%>\%
    purrr::map_dbl(1)
}
batch_it(x = as_year(2019), batches = month_lengths(2019))
}
