% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_poly.R
\name{bbox_to_poly}
\alias{bbox_to_poly}
\title{Convert a bbox to an sf geospatial polygon (rectangle)}
\usage{
bbox_to_poly(x, crs)
}
\arguments{
\item{x}{bbox - a named integer vector}

\item{crs}{Coordinate Reference System eg EPSG code}
}
\value{
an sf polygon object (sfc)
}
\description{
Convert a bbox to an sf geospatial polygon (rectangle)
}
