% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Wrapper to simplify download.file

Sometimes you just want to call download.file and not have to specify
the name of the destination file. You just want the local file to be the same
as the remote file that forms the basename of the URL.

Default setting for dir is "" which means file will end up in project root
directory (according to `here()`).

destfile we want to extract from the URL using `basename` but we need to offer
it as an argument to the function anyway because the user might want or need
to specify it, as you do with `download.file` eg if it turns out destfile
already exists.

User can specify `overwrite = TRUE` which will allow the file to be downloaded
even if the target already exists.}
\usage{
download_file(url, dir = "", destfile = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{url}{URL to download. Passed to `download.file()`.}

\item{dir}{(optional) Directory to download to.
Defaults to project root folder/working directory.}

\item{destfile}{(optional) Name of local file to download to.
The whole point of this is not to have to specify a destfile,
but if you want or need to, you can. Passed to `download.file()`.}

\item{overwrite}{(optional) Force allow overwrite of existing local destfile.
I haven't checked whether `download.file()` will actually do this in the end, though.}

\item{...}{Any other arguments you want to pass through to `download.file()`, such as
mode, method, headers, quiet and so on.}
}
\value{
None
}
\description{
Wrapper to simplify download.file

Sometimes you just want to call download.file and not have to specify
the name of the destination file. You just want the local file to be the same
as the remote file that forms the basename of the URL.

Default setting for dir is "" which means file will end up in project root
directory (according to `here()`).

destfile we want to extract from the URL using `basename` but we need to offer
it as an argument to the function anyway because the user might want or need
to specify it, as you do with `download.file` eg if it turns out destfile
already exists.

User can specify `overwrite = TRUE` which will allow the file to be downloaded
even if the target already exists.
}
\examples{
download_file("https://r-bootcamp.netlify.app/fishermen_mercury_README.md", dir = "data")

}
