% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_files}
\alias{download_files}
\title{Automatically download files from multiple URLs}
\usage{
download_files(urls, dir = here::here(), overwrite = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{urls}{A list or vector of URLs}

\item{dir}{(optional) Directory to download to. Defaults to the output of
\code{here::here()} i.e. the project directory or equivalent (since
\href{https://github.com/r-lib/rprojroot/pull/59}{this beauty} got merged you can
nest \code{here()}s and it just takes it in its stride.)}

\item{overwrite}{(optional) Force allow overwrite of existing local destfile.
I haven't checked whether \code{download.file()} will actually do this in the end,
though.}

\item{quiet}{don't say anything while downloading}

\item{...}{Any other arguments you want to pass through to \code{download.file()},
such as mode, method, headers and so on.}
}
\description{
Passes URLs to download_file() after checking that there are no duplicate
destination files
}
