% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji}
\alias{gitmoji}
\title{Writes emoji to the system clipboard, whose name(s) match \code{x}
Use \code{gitmoji_list()} to list all available gitmoji or to search the table of gitmoji by emoji name.
Use \code{gitmoji_desc()} to search the table of gitmoji by description}
\usage{
gitmoji(x, exact = TRUE)
}
\arguments{
\item{x}{string. A piece of text to search for}

\item{exact}{boolean, default TRUE. Whether to only return a result for the exact string provided, or return all names that contain the string.}
}
\value{
TRUE if successful; the emoji is/are copied to the clipboard
}
\description{
Writes emoji to the system clipboard, whose name(s) match \code{x}
Use \code{gitmoji_list()} to list all available gitmoji or to search the table of gitmoji by emoji name.
Use \code{gitmoji_desc()} to search the table of gitmoji by description
}
