% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitmoji.R
\name{gitmoji_desc}
\alias{gitmoji_desc}
\title{List gitmoji with a description matching \code{x}}
\usage{
gitmoji_desc(x = "")
}
\arguments{
\item{x}{string. A piece of text to search for.}
}
\value{
A three-column table showing matching emoji with their code
and description
}
\description{
If \code{x} is "" (as by default), all rows in the table will be returned
}
