% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmaps_to_sf.R
\name{gmaps_to_sf}
\alias{gmaps_to_sf}
\title{Convert Google Maps lat-lon Data to a Geospatial sf object}
\usage{
gmaps_to_sf(lst, crs)
}
\arguments{
\item{lst}{a list of coordinates as pasted from Google Maps right-click ($lat, $lon)}

\item{crs}{the desired CRS of the output}
}
\value{
an `sf` (sfc) object (collection)
}
\description{
If you right-click on a location in Google Maps, it gives you the option to copy the latitude and longitude. This is really useful. But it does provide the latitude figure first, which will be inaccurate if you try to feed it straight to `sf::st_point()`, which expects longitude first.
This function allows you to create a list of these coordinate pairs as pasted from GMaps, and convert it into a single, plottable, geospatial `sf` (sfc) object.
}
