% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_to_html5.R
\name{knit_to_html5}
\alias{knit_to_html5}
\title{Custom HTML5 knit function}
\usage{
knit_to_html5(input, quiet = FALSE, ...)
}
\arguments{
\item{input}{input document to be knitted}

\item{quiet}{passed to rmarkdown::render}

\item{...}{args passed to rmarkdown::html_document}
}
\description{
Custom HTML5 knit function
}
