% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrmidon-package.R
\docType{package}
\name{myrmidon}
\alias{myrmidon-package}
\alias{myrmidon}
\title{\code{myrmidon}}
\description{
"My R" Kitchen Drawer
}
\details{
Fran's personal useful tools, tricks and toys. Some functions I want to have
available most times I start up R, as well as my own HTML5 RMarkdown
template. Please feel free to use and improve them.
}
\keyword{internal}
