% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_along.R
\name{order_along}
\alias{order_along}
\title{Sort a data frame using a secondary key, within primary groupings}
\usage{
order_along(df, order_along, sort_by, desc = FALSE)
}
\arguments{
\item{desc}{}
}
\value{

}
\description{
Also works with `sf` tibbles – use `order_along.sf` instead.
}
