% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_data_join.R
\name{postcode_data_join}
\alias{postcode_data_join}
\title{Use postcodes.io To Get Postcode Metadata}
\usage{
postcode_data_join(df, var = postcode, fix_invalid = TRUE)
}
\arguments{
\item{df}{a data frame with a column of postcodes}
}
\description{
Use postcodes.io To Get Postcode Metadata
}
