% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_dates.R
\name{year_dates}
\alias{year_dates}
\title{Create vectors of dates for one or more years (and/or months)}
\usage{
year_dates(x, months = 1:12, unlist = FALSE)
}
\arguments{
\item{x}{a vector of years, provided as numeric or string}

\item{months}{a numeric vector. 1:12 by default (whole year).}

\item{unlist}{FALSE by default (allows returning a list). TRUE causes it to return a vector of dates.}
}
\value{
a vector of dates, or a list of vectors, one for each year, where multiple years are requested.
}
\description{
Create vectors of dates for one or more years (and/or months)
}
\examples{
head(year_dates(2019:2020, months = c(1, 4), unlist = TRUE), 10)
}
